rageshake: factor out submission to a separate file
This will mean we can load it asyncronously in future, if we want.
