+++ src/components/views/dialogs/BugReportDialog.js
@@ -16,7 +16,7 @@ limitations under the License.
 
 import React from 'react';
 import sdk from 'matrix-react-sdk';
-import rageshake from '../../../vector/rageshake';
+import submit_rageshake from '../../../vector/submit-rageshake';
 import SdkConfig from 'matrix-react-sdk/lib/SdkConfig';
 
 export default class BugReportDialog extends React.Component {
@@ -48,7 +48,7 @@ export default class BugReportDialog extends React.Component {
             return;
         }
         this.setState({ busy: true, err: null });
-        rageshake.sendBugReport(
+        submit_rageshake(
             SdkConfig.get().bug_report_endpoint_url, userText, sendLogs,
         ).then(() => {
             this.setState({ busy: false });
